Changes on values and increment type variable

+++ read_param.h
@@ -20,7 +20,7 @@ int seed;
 int conf_seed;
 double iRx, iRy, iRz; 
 double tmin, tmax, dt; 
-int increment; 
+double increment; 
 double accuracy; 
 double init_dir[3], dir1[3], dir2[3]; 
 bool surfdegen, uppersurf, lowersurf, DoubleU;
